requirements.txt
RPi.GPIO
fake_rpi
requests
rpi-rf